Spaces between Allow and Deny (in order) result in a 500 error.
Additionally, the order was wrong:
Deny,Allow will check allow even if deny was matched, and if allow matches too the final result is allow.
Allow,Deny will likewise check deny also after allow was matched, and deny the result if it matches the deny rule too. (Thus Allow,Deny is the proper order in a directory where we want to deny all.)